Rename safeway.py to app.py
